test(node): Add `mongoose` auto instrumentation test for `@sentry/node-experimental` (#10322)
This tests v5.13.22 of mongoose because it's the most commonly use
version currently supported by the OpenTelemetry plugin:

https://github.com/open-telemetry/opentelemetry-js-contrib/blob/fce7d3b5e478ff7525c9ffe99e59bf35f8c06207/plugins/node/instrumentation-mongoose/src/mongoose.ts#L80

https://www.npmjs.com/package/mongoose?activeTab=versions

It's worth noting that at least versions 5 and 7 of `mongoose` also
generate `mongodb` spans because they use the library internally and we
auto instrument everything. For some unknown reason, `mongoose@v8` does
not appear to generate `mongodb` spans and I haven't tested v6 because
very few people are using it.