test(node): Add `mysql2` auto instrumentation test for `@sentry/node-experimental` (#10259)
This PR adds a test for `mysql2` auto instrumentation for
`@sentry/node-experimental`.

`mysql2` will not query unless there is a connection and it does not
want to work without one like `mysql`.

This PR adds a `withDockerCompose` method to the test runner which
handles:
- Starting the docker container
- Waiting until some specific output has been seen from the container so
we know the server is up
- Closing and cleaning up the docker container/volumes

```ts
 createRunner(__dirname, 'scenario.js')
      .withDockerCompose({ workingDirectory: [__dirname], readyMatches: ['port: 3306'] })
      .expect({ transaction: EXPECTED_TRANSACTION })
      .start(done);
```

My only minor concern is that the mysql docker container creates a
volume to store data. If the cleanup code does not run, a 180MB volume
is left behind after every run. This will only be an issue when testing
locally but we could start to fill developers machines. These can be
cleaned up via `docker volume prune --force` but I would not want to run
this on peoples machines without telling them!