feat(tracing): Expose new `browserTracingIntegration` (#10351)
Extracted this out of
https://github.com/getsentry/sentry-javascript/pull/10327.

This PR:

* Introduces a new `browserTracingIntegration()`
* Does NOT deprecate BrowserTracing yet, as custom implementations in
Angular, Next, Sveltekit have not been migrated over yet, which would be
weird. We can deprecate it once we moved these over.
* Makes sure that custom implementations in Next & Sveltekit are "fixed"
automatically
* Uses a slim fork for the CDN bundles, to avoid shipping multiple
implementations in there.
* This means that in the CDN bundles, you can already use the new
syntax, but you cannot pass a custom routing instrumentation anymore,
and you also don't have the utility functions for it yet. I think this
is the best tradeoff for now, and it's probably not a super common case
to have custom routing instrumentation when using the Loader/CDN bundles
(and if you do, you have to stick to `new BrowserTracing()` until v8).

I copied the browser integration tests we have, which all pass!