feat(core): Deprecate `Span.setHttpStatus` in favor of `setHttpStatus` (#10268)
Deprecate the last remaining otel-incompatible API on the
`Span` class and interface - `Span.setHttpStatus`.

We replace this functionality with a `setHttpStatus` wrapper because in
addition to setting the data/tag value (more on that below), this
function also derives the span status from the status code. Added tests
to make sure this is covered properly (I think it wasn't covered well
before and we need to remove the old tests anyway).