test(e2e): Add Node SDK re-exports consistency e2e test (#10389)
Introduce an e2e test that checks if all
`@sentry/node` exports were re-exported in packages that depend on
`@sentry/node`. It's not a real e2e test, as there's no actual test
application, but we only execute a script that imports various SDKs
depending on `@sentry/node`. However, IMO it makes a lot of sense to
test the export of the final tarballs which we already have the
infrastructure for in our e2e tests.

see PR description for more details