ref(serverless): Convert integrations to functional approach (#10329)
This PR updates the serverless integrations to use the functional
approach, and rewrites all of the serverless unit tests.

In the unit test rewrite we remove the usage of global mocks and instead
do per file mocks. I dislike the amount of mocks that still exist, but
to prevent any regressions I only tried to translate the test code.