feat(core): Ensure `startSpan()` can handle spans that require parent (#10386)
This introduces a new `onlyIfParent` option to the start-span APIs,
which, if set, will ensure to not create a span if there is no parent.
I've added this also for otel to demonstrate this works there as well.

This PR also uses this option in a few places, esp. for http.client
spans, to ensure we do not capture these if there is no ongoing
transaction.

This replaces
https://github.com/getsentry/sentry-javascript/pull/10375/files with a
more generic solution, and includes the test from
https://github.com/getsentry/sentry-javascript/pull/10376 (which passes
here as well).