feat(node-experimental): Add koa integration (#10451)
This should auto-instrument Koa performance. We still needs tests for
this at some point. cc @timfish / @onurtemizkan

The main reason to add this is that we have docs for this on
docs.sentry.com, and we need a way to properly replace/update them (and
they currently rely a lot on `startTransaction()` etc.) With this we can
eventually just update this to use the default koa integration we
provide.