build(ci): Ensure we run E2E tests when profiling node is skipped (#10512)
With the recent change to skip the node profiling compile step when it
was not changed, we implicitly also skipped the "Prepare E2E tests" job
- as that is by default skipped if the dependent job is skipped.

This PR changes this so that we should be running this even if that was
skipped. We have to make sure to check that `build` itself was not
skipped though (as then we want to _actually_ skip this).