build(profiling-node): make sure debug build plugin is used (#10534)
Fixes https://github.com/getsentry/sentry-javascript/issues/10525

When writing the rollup config, we didn't include the debug build
plugin. This led to things not be replaced properly as `profiling-node`
bundles everything into a single file.

This was also causing issues in our CI:
https://github.com/getsentry/sentry-javascript/actions/runs/7804351046/job/21287026518?pr=10527

Backporting this fix to v7 so we can do a `7.100.1` release after we
merge this in.