feat(core): Streamline custom hub creation for node-experimental (#10555)
This is an intermediate step to streamline what we need to customize in
OTEL. This should make it easier to slowly remove customizations there.

The main change is that you can set a `createHub` method on the main
carrier, which will be used if set. This way, all the code we have to
duplicate in node-experimental right now just to create the correct
`OpenTelemetryHub` can mostly go away.

We will remove this again during then v8 cycle.