ref(replay): Use `beforeAddBreadcrumb` hook instead of scope listener (#10600)
Currently, replay captures breadcrumbs by setting up a scope listener on
the current scope. However, this is flawed because this means it will
not pick up breadcrumbs that are added to the isolation scope (which
will be the default going forward, in v8).

This PR changes this to use the `beforeAddBreadcrumb` hook instead to
listen to this.

CAVEATS: This means we only capture breadcrumbs added via
`Sentry.addBreadcrumb()`. If somebody adds a breadcrumb directly to a
scope, the hook is not triggered. I _think_ this is OK, but something to
be aware of - if we want to change this, we'd need to trigger the hook
on the scope, which may be a bit iffy/tricky. But for the most part,
this shouldn't matter, hopefully - this is also non-critical, meaning if
we miss a breadcrumb (for whatever reason) it's not _that_ bad, it will
simply not be visible in the Replay.