ref(core): Make `on` and `emit` required on client (#10603)
This allows us to remove a bunch of code around fallbacks etc. Esp. in
replay this also allows us to remove a bunch of code around breadcrumbs
etc.

note @JonasBa I did not remove this in profiling-node yet because the
test rely on this quite a bunch, can you take a look at this when you've
got some time?