feat(core): Lookup client on current scope, not hub (#10635)
One more step to avoid the hub...

Note: This showed a potential footgun (?), where if you pass a
client-less scope to `startSpan()`, no span will be created (because no
client = no tracing, and the given scope will be made the current scope
in the callback). this is correct behavior here, but _could_ be a bit
unexpected ("why is no span being created here??"). We could think about
warning if a client-less scope is passed in there, but not sure if
that's worth it...