feat: Implement new Async Context Strategy (#10647)
This updates the ACS to not rely on hubs (only) anymore.

Now, instead of the ACS providing only `getCurrentHub` and
`runWithAsyncContext`, this is changed a bit:

1. There is always a strategy, even if running in browser or similar. we
just use the default (=stack) strategy in that case.
2. The ACS always returns a hub/scope/etc, not `undefined` - so the
strategy must take care of falling back to the global hub itself (to
make it easier to implement hub<>scope interop).

The ACS defines the following methods now:

* `getCurrentScope`
* `getIsolationScope`
* `withScope`
* `withSetScope` - a variant that takes a scope and makes it the active
one. I decided to make this a dedicated method on the ACS instead of
overloading `withScope` because the types for that are rather tricky to
repeat in strategies...
* `withIsolationScope`
* `withSetIsolationScope` - not we do not use this yet, but to keep the
door open for us this is already required.
* `getCurrentHub` --> for now, for backwards compatibility

The methods themselves are pretty straightforward now! We basically
always create a new hub, and decide if we want to fork the
current/isolation scope based on what method has been used. We still use
a hub everywhere under the hood for now.