feat(core): Decouple metrics aggregation from client (#10628)
This is a re-opening of #10596 which was closed automatically after the
branch is was based off was closed.

Part of the work for #10595

This PR decouples metrics aggregation from the client and stores the
aggregator globally the first time it is required.

This PR does not remove `captureAggregateMetrics` from the `BaseClient`
or make `sendEnvelope` public as this can occur in a later PR. This will
remove the remaining metrics code from the most basic tree-shaken Sentry
configuration.

The default (node) metrics implementation remains in `@sentry/core`
because it's required for Deno and Vercel-Edge which don't ref.
`@sentry/node`.

There's a load of extra deletions and cleanup that can occur for v8 when
the integration is deleted. For example `BrowserMetricsAggregator` can
probably move to `@sentry/browser`.