fx(node): Fix anr worker check (#10719)
I've removed the `hub` global on the carrier in a recent PR, and noticed
this leftover. Guess this is not used/covered by tests...? It should
work with using `acs` as that should now always be set too (also for the
default case), unless sentry is not initialized. But for type safety,
the acs _may_ be undefined.