feat(opentelemetry): Align span options with core span options (#10761)
This aligns the options for `startSpan` in otel with the core ones. Only
the `kind` is there in addition for now (we may want to remove this,
let's see).

Especially, this also adds the ability to pass a `scope` there and pick
up the correct context & span to continue, as well as adding tests for
the options.