feat(opentelemetry): Ensure DSC & attributes are correctly set (#10806)
This PR fixes a few things about the OTEL handling of our SDK.

1. Ensure that attributes are actually correctly put on the trace data -
we used to remove e.g. `sentry.sample_rate` which we don't do in "main"
SDK. Question: Should we remove the `otel.attributes` context data then?
As this is kind of redundant at this point with the `data` on the trace?
2. Ensure we put the `transaction` tag correctly on the transaction, to
align with "main" SDK.
3. Ensure we store the incoming DSC as traceState, so we can pick this
up. This can include e.g. transaction name and stuff like this to
persist.