fix(processing): Only mark aggregate errors as exception groups (#10850)
When Sentry started supporting the idea of exception groups, two changes
happened. In the SDK, we adapted our logic for handling linked errors to
also handle `AggregateError`s. And in our ingest pipeline, we began
looking for an `is_exception_group` flag on the last entry in
`event.exception.values; when we found it, we'd then ignore that entry
when grouping and titling events, under the assumption that it was just
a container and therefore wasn't meaningful.

When it came to instances of `AggregateError`, this worked great. For
linked errors, however, this caused us to focus on the `cause` error
rather than the error which was actually caught, with the result that it
both threw off grouping and made for some very unhelpful titling of
issues. (See the screenshot below, in which the first three errors are,
respectively, an `UndefinedResponseBodyError`, a `RequestError`, and an
`InternalServerError`, though you'd be hard pressed to figure that out
without opening them up.)

This fixes those problems by restricting the use of the
`is_exception_group` flag to `AggregateError`s.

Note: In order to update the tests to work with this change, I had add
in consideration of the error `name` property and the corresponding
event `type` property, to match what we do in real life. To keep things
readable, there's a new mock `AggregateError` class, which I adapted all
the tests to use.