feat(core): Remove deprecated props from `Span` interface (#10854)
Instead, in places we need it we cast to a `SentrySpan` which still has
the things in place, for now.

With this, our span interface is _almost_ the same as for otel spans -
missing are only:
* Aligning `traceFlags` - this has apparently been updated in OTEL to
have type `number` only, which makes this a bit easier.
* Aligning `setStatus` which has a different signature in OTEL.

I'll do these in follow ups!

The biggest work here was fixing tests - I tried to rewrite tests to do
less mocking where possible, which IMHO should cover actual
functionality better than before (e.g. in svelte).