feat(profiling-node): Expose `nodeProfilingIntegration` (#10868)
@JonasBa when you got some time, could you look into actually rewriting
this to the new style? Right now a lot of the tests depend on the
non-hook based version etc. and I'm not 100% clear on what can/cannot be
removed there 😅

The end result should be just having a single `nodeProfilingIntegration`
that receives the client in `setup(client)` and sets up hooks etc in
there!