feat(esm): Import rather than require `inspector` (#10910)
This does break support for custom builds of Node which have been built
without inspector support. The only one that comes to mind is Vercel/pkg
but this has recently been deprecated since Node v21 supports building
single file executables.

We may come back later and find a way to work around this but we cannot
use async `import` with the current code, so we'll need to revisit this
after we have a proper ESM build working!

This also removes `local-variables-async.ts` since it's not used and
async inspector requires a worker thread to work.