fix(types): Add `addScopeListener` to `Scope` interface (#10952)
`getCurrentScope()` has been changed to return the `Scope` interface
rather than the `Scope` class. This means it's no longer possible to
access `addScopeListener` on the returned scope.

This PR adds this method to the interface. I also updated all the jsdocs
in the class to use `@inheritDoc` so these docs aren't duplicated.