feat(remix): Add Vite dev-mode support to Express instrumentation (v7). (#10811)
Adds dev-mode support to Remix setups with Vite and Express.

We currently accept Remix server `build` as an object to instrument. But Remix allows `build` as a synchronous or asynchronous function that returns the build object. Currently, it seems that functions are only used in development servers, and not in production. So, while this update slightly reduces `requestHandler` performance on dev servers, it does not on production builds.

We need `build` in 2 places:

1- We instrument the loaders / actions on build, then we pass them down to the original implementations.
2- We use the `routes` inside them to create parameterised transactions.

This update adds new internal wrappers around them to make sure that we don't miss out on the returned / resolved values in case `build` is a function, for both cases. 

This PR also adds a new E2E test application using the latest Remix version and Vite, and it runs the tests on `dev` mode.