feat(browser): Use idle span for browser tracing (#10957)
And remove idle transaction.

With this, we can afterwards remove the span recorder, and everything
related to it 🎉

I tried to keep all functionality intact. Some small cleanups I made:

- idle span finish reason is kept as an attribute - to keep backwards
compat I finally also write it as a tag on the transaction event
- idle span trimming happens automatically now (no more `trimEnd` option
- we always set this to true anyhow). However, this only applies when
the span is auto-ended by the idle functionality, not if manually
calling `span.end()` - which is fine IMHO because that will usually not
happen for users, and even so it's not the end of the world.
- We now use `continueTrace` for trace propagation of the pageload span.
- no more `location` on the sampling context - we talked about this
before, this is just gone, and users can instead either access the
attributes from the sampling context, or just do `window.location.href`
themselves, which should have the same outcome.