feat(sveltekit): Export `unstable_sentryVitePluginOptions` for full Vite plugin customization (#10930)
We decided to provide an escape hatch to the reduced (but therefore 
version-agnostic) Vite plugin option API in the SvelteKit version of the 
plugin (`sentrySvelteKit()`). By specifying options in `unstable_vitePluginOptions`, 
users can still set and override all options the Vite plugin can take. 
We explicitly warn that keys in `unstable_vitePluginOptions` do not 
follow semver and might be removed or changed at any time.