fix(react): Set `handled` value in ErrorBoundary depending on fallback (#10989)
Previously, we made a change to always mark such errors as unhandled
(#8914). However, as pointed out and discussed in #10985, this isn't
always the best default value. I propose we set handled/unhandled
depending on the presence of the `fallback` component. If users specify
a fallback, I think it's safe to mark the error handled, given that
users show something else instead of the errored component.