ref(nextjs): Clean up browser tracing integration (#11022)
Updates the app & pages router instrumentation to avoid depreacted
options & streamline it overall.

This required a change to browser tracing, which is to allow to provide
a custom `sentryTrace` / `baggage` string value to continue a trace
from. Before this, we always picked up the trace from `meta` tag, which
meant that if we tried to continue a trace manually in nextjs, then
called the default browser tracing integration, it would just overwrite
it.

Now, you can optionally pass this like this:

```js
startBrowserTracingPageLoadSpan(
    client,
    { name },
    { sentryTrace, baggage }
});
```

And this will take presedence over looking at the meta tags! This also
means we don't need any custom scope/trace handling in nextjs there
anymore, which is also nice!