feat(v8): Add @sentry/aws-serverless package (#11052)
This change renames the `@sentry/serverless` SDK to be
`@sentry/aws-serverless` and removes all GCP related code. It also adds
a migration note detailing the changes we made.

It also updates the package to only emit CJS, because it relies on
monkeypatching via `require` calls. We can add back ESM at a later
point, will create follow up issue for it.

This PR does not touch any of the lambda layer code, I will come back
and refactor it afterwards (too scary to think about right now)

We will rename the `@sentry/google-cloud` package to
`@sentry/google-cloud-serverless` in a follow up PR