ci: Re-add size limit action (#11096)
Size seems to be much increased, let's figure this out in follow ups.

Also there is some issue with it reading next.js sizes, also something
we can look at in follow ups... I disabled these size checks for now.

The action is now inlined as a package in the monorepo, as our fork has
been archived and the upstream package is not really maintained and
works pretty different from what we need. Now we at least have full
control over this.