feat(v8/nextjs): Remove usage of class integrations (#11182)
In order to proceed with removing `Sentry.Integrations.X` as per
https://github.com/getsentry/sentry-javascript/issues/8844, there's
still some places to clean up.

This does conflict with
https://github.com/getsentry/sentry-javascript/pull/11016, but not sure
when that merges in, so opening this in the meantime to unblock the
integrations cleanup work. if we think the OTEL nextjs work will merge
in sooner then the next 1-2 days, I'm happy to leave this alone for now!