fix(metrics): use web-vitals ttfb calculation (#11185)
Recommend reading through https://web.dev/articles/ttfb before review.

In
https://www.notion.so/sentry/TTFB-vital-is-0-for-navigation-events-2337114dd75542569eb70255a467aba6
we identified that ttfb was being incorrectly calculated in certain
scenarios. This was because we were calculating `ttfb` relative to
transaction start time, **before** it has been adjusted by `browser`
related spans about request/response ( remember browser tracing adjusts
the start timestamp of a pageload transaction after adding certain
request/response related spans).

This meant that `Math.max(responseStartTimestamp - transactionStartTime,
0)` would just end up being `0` most of the time because using
`transactionStartTime` was not correct.

To fix this, we avoid trying to rely on our `transactionStartTime`
timestamp at all, but instead using the web vitals version helper for
this.

When this gets merged in, I'll backport it to v7. I'm doing this in v8
first because I don't want to deal with the merge conflict that comes
when we eventually migrate this code from tracing internal into the
browser package.

As a next step, we should seriously think about getting rid of all of
our vendored code and just rely on the web vitals library - it's a huge
pain to maintain this, and I'm sure there are some insidious bugs
sneaking about.