fix(sveltekit): Ensure sub requests are recorded as child spans of parent request (#11130)
When switching the SvelteKit server side SDK to `@sentry/node` powered
by Otel, the semantics behind `continueTrace` changed as outlined in
#11199. TLDR: We previously called `continueTrace` in a nested way when
dealing with sub-requests`*` in SvelteKit. In our old Node SDK, this did
nothing; in the new SDK, this currently causes a new root
span/transaction to be created for the sub request.

This patch now ensures that we continue to send sub request spans as child
spans of the top-level/parent request.