fix(node): Local variables skipped after Promise (#11234)
The debugger call stack does not include the `new Promise` frames that
are parsed from `error.stack`. This means that when we go through the
frames to apply the local variables, the frames don't match up and we
bail.

This patch ignores those frames when matching functions in the frames.