feat(node): Simplify `SentrySpanProcessor` (#11273)
Instead of extending `BatchSpanProcessor`, we handle this ourselves now.

There may be additional improvements we can do here, but this should be
OK for now!

I added two tests to cover what we really want to achieve/show here:

1. Ensure that if a span & child span are ended after each other, but in
the same tick, they are correctly sent.
2. Ensure that if a child span is ended later, it is skipped.

By skipping the batched span processor, we can simplify our flushing a
bit and also get rid of some of the delays.

For now I kept the exporter as a separate class (it is not exported
anyhow, so purely internal), we can eventually look into merging this or
not, but this is just a moving-parts-around exercise then. This way,
tests continued to work mostly, which is good I'd say.