feat(utils): Allow text encoder/decoder polyfill from global __SENTRY__ (#11283)
The TextEncoder option was removed in
https://github.com/getsentry/sentry-javascript/pull/10701 because all of
the newly supported platforms support TextEncoder/Decoder.

Sadly that's not true for React Native, yet. TextEncoder will be
included in the next release of RN with Hermes. See the PR
https://github.com/facebook/hermes/commit/3863a36a53005dd1e6d39ea0d4ef5573bafde910#diff-4bf4a4ab271f254baf2097be87be98333ec69eb2d41e074b81147656c33145eb

We can not drop support of all the previous RN versions in the RN SDK v6
(the next major).

To avoid passing the encoder option around the SDK, I propose adding
polyfill properties to the global __SENTRY__ object which RN can use.