feat(core): Handle string sample rates
If users pass e.g. `sampleRate: process.env.SAMPLE_RATE`, this will be somewhat silently ignored because we ignore non-number sample rates. This is a bit of a footgun.

So this PR changes this so that we actually accept such sample rates.
In the process, I also removed the `isNaN` polyfill as this is not really needed anymore.