feat(v8/gatsby): Update SDK initialization for gatsby (#11292)
This is quite the refactor, I recommend reading through the migration
docs to understand what the user facing API should be.

1. Remove deprecated SDK init via options

We used to initialize the SDK via Gatsby plugin options in the Gatsby
SDK. This is very restrictive though, and due to how Gatsby serializes
it's options can lead to bugs. Later on in v7 of the SDK we introduced a
`sentry.config.js` approach
([docs](https://docs.sentry.io/platforms/javascript/guides/gatsby/)).
This PR removes support for the plugin option deprecation.

- Remove automatic adding of `browserTracingIntegration`

Because of the plugin option initialization, we used to add
`browserTracingIntegration` integration automatically via SDK init. v8
removes this because of bundle size + we don't have good routing
instrumentation for gatsby.

- Upgrade gatsby webpack plugin to 2.0, and add option to configure it's
usage

resolves https://github.com/getsentry/sentry-javascript/issues/9837 - I
called the option to control this `enableClientWebpackPlugin`.