feat(feedback/v7): Make "required" text for input elements configurable (#11287)
Backport of https://github.com/getsentry/sentry-javascript/pull/11153

From the OP (h/t @soerface):

This introduces an option `isRequiredLabel`, and I planned to also add
`errorMessageText`.

However, this PR is branched off from a version that is a couple days
old. Right now, @c298lee is currently working on getsentry/sentry#63749,
which is causing major changes and the current version on master doesn't
work yet. Therefore, I didn't yet implement `errorMessageText`.

So consider this PR as a PoC, and either feel free to tag me when the
screenshot changes are done – then I'll redo the changes based on the
version that supports screenshots – or add the option on your own;
however you prefer :slightly_smiling_face:

---------

Co-authored-by: Soeren Wegener <wegener92@gmail.com>