fix(node): Fix baggage propagation (#11363)
The node integration tests (correctly!) showed that we were not
propagating baggage 100% correct anymore. Turns out there were a few
issues with what we did before:

## Do not freeze DSC onto span `traceState` when not continuing incoming
trace

The main issue was, that we did not actually set the `transaction`,
`sampled` and `sample_rate` baggage entries. This was the case because
we only picked these from the trace state, but we only set this fully in
our own `startSpan()` APIs, not when auto instrumentation created spans.
So I changed this to _not_ set the trace state in `startSpan()` APIs,
but instead leave it empty _unless_ we are actually continuing an
incoming trace. Whenever we are _not_ continuing an incoming trace,
we'll now generate the DSC at injection-time based of the actual span
data.

So to summarize, we now set the `dsc` trace state in two places:

1. In the propagator, when we `extract` an incoming trace.
2. In our trace functions, when we use `forceTransaction` which we treat
as continuing an incoming trace.

## No merging of existing baggage headers

Another issue that came up is that we did not properly handled other
`baggage` entires being set. This is because of how OTEL does injection,
where it passes the `carrier` as object which we treated as `unknown`
and basically ignored. This meant that if a user did e.g. `fetch({
headers: { baggage: 'my=value' }})`, we would overwrite the baggage and
not keep the exiting value(s). This PR fixes this as well by ensuring we
keep this around.

## Future improvements

After talking with @lforst, we may be able to improve/streamline this
further by getting rid of the DSC trace state.
The logic could be:

1. If there is a DSC on the scope's propagation context, use that.
2. Else, use the root span

We may update to this logic internally in a follow up step!