ref(feedback): Create stub integrations for feedback modal & screenshot integrations (#11342)
We're planning to split up the modal and screenshot functionality of the
feedback widget into their own integrations, so they can be loaded async
if folks want to do that. Loading async would have the benefit that the
code is only needed when the user interacts with the feedback widget,
and we're actually going to render the modal whether that be with or
without the screenshot input/editor.

This PR sets up some stubs for where the code will live. There's a bunch
of boilerplate for the new integrations inside packages/* and also lots
of exports to setup so that the stubs are available to be imported into
apps.

Currently if someone wanted to setup feedback with screenshots they
would be doing:
```javascript
import * as Sentry from "@sentry/browser";
import {feedbackIntegration, feedbackModalIntegration, feedbackScreenshotIntegration} from '@sentry-internal/feedback';

...
  integrations: [
    feedbackIntegration(),
    feedbackModalIntegration(),
    feedbackScreenshotIntegration(),
  ]
```

**After this PR people will keep doing that.**

In a followup PR we'll move the implementation from
`@sentry-internal/feedback` into the new
`@sentry-internal/feedback-modal` and
`@sentry-internal/feedback-screenshot`, so people will be able to do
this instead:
```diff
  import * as Sentry from "@sentry/browser";
- import {feedbackIntegration, feedbackModalIntegration, feedbackScreenshotIntegration} from '@sentry-internal/feedback';
+ import {feedbackIntegration} from '@sentry-internal/feedback';
+ import {feedbackModalIntegration} from '@sentry-internal/feedback-modal';
+ import {feedbackScreenshotIntegration} from '@sentry-internal/feedback-screenshot';

  ...
    integrations: [
      feedbackIntegration(),
      feedbackModalIntegration(),
      feedbackScreenshotIntegration(),
    ]
```
Equally valid, is people will be able to import everything from
`@sentry/browser` too, ie `import * as Sentry from "@sentry/browser";`
with `Sentry.feedbackScreenshotIntegration()`

Related to https://github.com/getsentry/sentry-javascript/issues/11435