feat(core): Implement `suppressTracing` (#11468)
This implements a new `suppressTracing` method which can be used to
ensure we do not send any spans inside of a callback.

Usage:

```js
await suppressTracing(() => {
  // do not capture any spans for this fetch!
  return fetch(...);
});
```

This also changes the behavior of starting child spans slightly to
actually create `NonRecordingSpan` as child spans when not sampled
(previously, we'd still create SentrySpans). I choose to continue to
create proper spans for unsampled transactions because it makes things
like picking name and stuff like this easier for the time being.